[frontend] refacto files FilterInputs
